// src/store/API.interface.ts
export interface TokenItem {
    title: string
    type: string
    properties: {
        name: {
            type: string
            description: string
        }
        description: {
            type: string
            description: string
        }
        image: {
            type: string
            description: string
        }
    }
}